Moved Oscillator to format common utils
