+++ internal/format/s3m/playback/effect/effect_settremolowaveform.go
@@ -3,7 +3,8 @@ package effect
 import (
 	"fmt"
 
-	"gotracker/internal/format/s3m/layout/channel"
+	formatutil "gotracker/internal/format/internal/util"
+	"gotracker/internal/format/xm/layout/channel"
 	"gotracker/internal/player/intf"
 )
 
@@ -18,7 +19,7 @@ func (e SetTremoloWaveform) Start(cs intf.Channel, p intf.Playback) {
 
 	mem := cs.GetMemory().(*channel.Memory)
 	trem := mem.TremoloOscillator()
-	trem.Table = channel.WaveTableSelect(x)
+	trem.Table = formatutil.WaveTableSelect(x)
 }
 
 func (e SetTremoloWaveform) String() string {
