+++ internal/format/xm/layout/channel/memory.go
@@ -1,5 +1,9 @@
 package channel
 
+import (
+	formatutil "gotracker/internal/format/internal/util"
+)
+
 // Memory is the storage object for custom effect/effect values
 type Memory struct {
 	portaToNote         uint8
@@ -25,8 +29,8 @@ type Memory struct {
 	LinearFreqSlides bool
 
 	tremorMem         Tremor
-	vibratoOscillator Oscillator
-	tremoloOscillator Oscillator
+	vibratoOscillator formatutil.Oscillator
+	tremoloOscillator formatutil.Oscillator
 	patternLoop       PatternLoop
 }
 
@@ -136,18 +140,18 @@ func (m *Memory) TremorMem() *Tremor {
 }
 
 // VibratoOscillator returns the Vibrato oscillator object
-func (m *Memory) VibratoOscillator() *Oscillator {
+func (m *Memory) VibratoOscillator() *formatutil.Oscillator {
 	return &m.vibratoOscillator
 }
 
 // TremoloOscillator returns the Tremolo oscillator object
-func (m *Memory) TremoloOscillator() *Oscillator {
+func (m *Memory) TremoloOscillator() *formatutil.Oscillator {
 	return &m.tremoloOscillator
 }
 
 // Retrigger runs certain operations when a note is retriggered
 func (m *Memory) Retrigger() {
-	for _, osc := range []*Oscillator{m.VibratoOscillator(), m.TremoloOscillator()} {
+	for _, osc := range []*formatutil.Oscillator{m.VibratoOscillator(), m.TremoloOscillator()} {
 		osc.Reset()
 	}
 }
