+++ internal/format/xm/playback/effect/util.go
@@ -1,6 +1,7 @@
 package effect
 
 import (
+	formatutil "gotracker/internal/format/internal/util"
 	"gotracker/internal/format/xm/layout/channel"
 	"gotracker/internal/format/xm/playback/util"
 	"gotracker/internal/player/intf"
@@ -166,7 +167,7 @@ func doTremolo(cs intf.Channel, currentTick int, speed uint8, depth uint8, multi
 	doVolSlide(cs, delta, 1.0)
 }
 
-func calculateWaveTable(cs intf.Channel, currentTick int, speed uint8, depth uint8, multiplier float32, o *channel.Oscillator) float32 {
+func calculateWaveTable(cs intf.Channel, currentTick int, speed uint8, depth uint8, multiplier float32, o *formatutil.Oscillator) float32 {
 	delta := o.GetWave(float32(depth) * multiplier)
 	o.Advance(int(speed))
 	return delta
